Refactor pr suggestions
